Fix panda build.
The recovery UI has changed in the main project bootable/recovery.  It is
now a C++ class that each platform should inherit and make device specific.

I have updated file recovery_ui.cpp to use this class based on other devices.
I have not tested that the recovery UI or the boot loader works. This change
is only to fix the build.  The result of this build should not be used without
testing first.

Change-Id: Iaca6c06cca4bdaff64d51ab0379dbcb11ff9bf47