Add dummy entries for /boot and /recovery to unified fstab
The original recovery.fstab file for panda had some dummy entries
that clearly were incorrect, so I did not add them to the
unified fstab.  However, the ota tools expect there to be
at least a /boot and /recovery partition, so this fix just copies
all the dummy entries from the old recovery fstab to the unified
fstab to fix the build.

Change-Id: I8a89541e6ccd494175c22de24040334961313250